Finalized the cache structure.
This object cache provides an interface for allocating and freeing memory
blocks more efficiently (in some cases). Instead of explicitly freeing each
memory block when it is not needed, the block is cached for reuse.  When the
same sized block is requested in the future a preallocated block will be
returned. This may decrease the load on the operating system, especially in
multithreaded applications.

An array based queue is maintained internally to keep track of the cached
objects.